pyproject.toml
[tool.poetry]
name = "pythonproject11"
version = "0.1.0"
description = ""
authors = ["Oksana Konkina <oksanka.808@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.5"
beautifulsoup4 = "^4.14.2"
psycopg2-binary = "^2.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
